requirements.txt
amqp==2.6.1
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Django==3.1.6
django-analytical==3.0.0
django-analytics-client==0.3.0
django-crispy-forms==1.11.0
django-google-analytics-app==5.0.2
django-tracking==0.2.7
django-user-tracking==0.4.11
django-user-visit==0.4.1
httpagentparser==1.9.1
idna==2.10
kombu==4.6.11
Pillow==8.1.0
pycparser==2.20
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
structlog==20.2.0
ua-parser==0.10.0
urllib3==1.26.3
user-agents==2.2.0
vine==1.3.0
